language_doc/pyproject.toml
[project]
name = "autonote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.14.0",
    "pillow>=12.0.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
]
